requirements.txt
argparse == 1.4.0
psycopg2 == 2.9.7
python-dotenv
flask
bs4
requests
json
csv
numpy
torch
torchvision
pandas
random